requirements.txt
streamlit
faiss-cpu
sentence-transformers
transformers
torch
beautifulsoup4
requests
numpy
